Gave RECEPTACLE its own CNTFCN to handle burning objects.
Though this doesn't work as well as I had hoped, since it can only be
used to prevent explicit taking, e.g. "TAKE NEWSPAPER". It won't prevent
implicit taking, e.g. "READ NEWSPAPER".

So BALLOON-BURN still has to clear TAKEBIT and READBIT for things to
work, which means you won't be able to take (or read) the object even if
extinguish it.