Worked around problem with CNTFCN.
Since CNTFCN can't be used to precent implicit taking, give TRYTAKEBIT
to any object being burned in the receptacle.

That means that if you then extinguish the object, it won't be
implicitly taken by e.g. "READ NEWSPAPER". I'm sure that issue could be
worked around as well, but I think it's still an improvement.

You could argue that the newspaper should no longer be readable after
it's been burning, but it's never described as a "party burnt newspaper"
or anything like that, so I'll attribute that to Zorkian magic.