Changed SERPENT's and PALANTIR-3's LDESC to FDESC
Both of these objects are inside containers, so their LDESC is never
printed. FDESC, on the other hand, is so that's probably what the author
intended.

In the case of PALANTIR-3, there was a hack to clear LDESC once it was
picked up, so it even acted like an FDESC. That hack is, of course, no
longer needed.